Fix GetAuthRequestUrl code.
