+++ src/client/Microsoft.Identity.Client.Services/Internal/Requests/Interactive/IAuthCodeRequestComponent.cs
