+++ src/client/Microsoft.Identity.Client.Services/UI/AuthorizationResult.cs
