+++ src/client/Microsoft.Identity.Client/ApiConfig/Parameters/AcquireTokenInteractiveParameters.cs
@@ -12,23 +12,18 @@ using Microsoft.Identity.Client.Utils;
 namespace Microsoft.Identity.Client.ApiConfig.Parameters
 {
 
-    internal class AcquireTokenInteractiveParameters : IAcquireTokenParameters
+    internal class AcquireTokenInteractiveParameters : GetAuthorizationRequestUrlParameters, IAcquireTokenParameters
     {
-        public Prompt Prompt { get; set; } = Prompt.NotSpecified;
         public CoreUIParent UiParent { get; } = new CoreUIParent();
 
         /// <summary>
         ///  These need to be asked for to the /authorize endpoint (for consent)
         ///  but not to the /token endpoint
         /// </summary>
-        public IEnumerable<string> ExtraScopesToConsent { get; set; } = CollectionHelpers.GetEmptyReadOnlyList<string>();
         public WebViewPreference UseEmbeddedWebView { get; set; } = WebViewPreference.NotSpecified;
-        public string LoginHint { get; set; }
-        public IAccount Account { get; set; }
         public ICustomWebUi CustomWebUi { get; set; }
-        public string CodeVerifier { get; set; }
 
-        public void LogParameters(ILoggerAdapter logger)
+        public new void LogParameters(ILoggerAdapter logger)
         {
             var builder = new StringBuilder();
             builder.AppendLine("=== InteractiveParameters Data ===");
