+++ src/client/Microsoft.Identity.Client/ApiConfig/Parameters/WebViewPreference.cs
